backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
bcrypt==3.2.0
slowapi==0.1.9
cohere>=5.0.0
python-dateutil==2.8.2